rename `models.py` to `types_.py`
