+++ .agents/skills/mkdocs-generation/templates/mkdocs.yml
@@ -59,4 +59,4 @@ nav:
 #       - Overview: api/index.md
 #       - Core:
 #           - Client: api/client.md
-#           - Models: api/models.md
+#           - Types: api/types_.md
